Add my content
In a separate github repo I started developing my website. I used .html, .js, and .css but not react. It is important to me that my website is responsive to any screen. So this is the start of moving my content over and modifying each section to consider smooth interactions