Updates
It's been a minute so it's just a lot of updates including picture, publications, and wanting to refine the project section images and pop up. I also temporarily removed the gallery section. I want to mess around more with which photos I show and the layout of the pop up